Combine all regularizer tests.
